fix: update document
